fix(mcp): avoid cloud exception import cycle
